Split WhatsApp webhook orchestration into focused modules
