+++ server/_core/whatsappTypes.ts
@@ -1,0 +1,11 @@
+import type { Lang } from "./i18n";
+import type { NormalizedInboundMessage } from "./normalizedInboundMessage";
+
+export type NormalizedWhatsAppEvent = NormalizedInboundMessage & {
+  channel: "whatsapp";
+};
+
+export type WhatsAppHandlerContext = {
+  reqId: string;
+  lang: Lang;
+};
